# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Essa é a API de Controle de Caixa, que pode ser utilizada junto com a aplicação para controle de caixa de pequenos négocios",
        "version": "1.0.1",
        "title": "Controle de Caixa API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "carleonepsouza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://app-loja.azurewebsites.net/v1/api"
        }
    ],
    "host": "https://app-loja.azurewebsites.net/v1/api",
...
    "basePath": "/v1",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/products": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Register Product",
                "description": "This route will be responsible for register a product",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Cerveja Devassa",
                                        "description": "Ceverja Lata",
                                        "price": 8,
                                        "classification": "ML",
                                        "category": "6408c5fca1bb1f33b014dee2",
                                        "volume": "350",
                                        "quantity": 100,
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": " Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "409": {
                        "description": "Conflit!"
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Return a list of products",
                "description": "This route will be responsible for return a list of products",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Return a specific product",
                "description": "This route will be responsible for return a product by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Update a specific product",
                "description": "This route will be responsible for update a product by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Cerveja Devassa",
                                        "description": "Ceverja Lata",
                                        "price": 8,
                                        "classification": "ML",
                                        "category": {
                                            "types": "object",
                                            "$ref": "#/components/schemas/Category"
                                        },
                                        "volume": "350",
                                        "quantity": 100,
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": " Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Delete a specific product",
                "description": "This route will be responsible for delete a product by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "value": {
                        "type": "number"
                    },
                    "classification": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "number"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}